fp32 loss accumulator on top of bf16 forward autocast
Wrap only the model forward in torch.autocast(bf16); cast pred back to
fp32 before the masked sum / L1 loss accumulation. Matmul-heavy compute
stays in bf16 (preserves the ~24% per-epoch speedup), loss accumulation
runs in fp32 to remove the precision noise on small residuals at the
cosine tail.

+++ train.py
@@ -374,7 +374,7 @@ class Config:
     agent: str | None = None
     debug: bool = False
     skip_test: bool = False  # skip final test evaluation
-    amp_bf16: bool = True  # use bfloat16 autocast for training forward + loss
+    amp_bf16: bool = True  # use bfloat16 autocast for model forward; loss accumulator stays in fp32
 
 
 cfg = sp.parse(Config)
@@ -469,12 +469,14 @@ for epoch in range(MAX_EPOCHS):
 
         with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=cfg.amp_bf16):
             pred = model({"x": x_norm})["preds"]
-            abs_err = (pred - y_norm).abs()
-            vol_mask = mask & ~is_surface
-            surf_mask = mask & is_surface
-            vol_loss = (abs_err * vol_mask.unsqueeze(-1)).sum() / vol_mask.sum().clamp(min=1)
-            surf_loss = (abs_err * surf_mask.unsqueeze(-1)).sum() / surf_mask.sum().clamp(min=1)
-            loss = vol_loss + cfg.surf_weight * surf_loss
+
+        pred = pred.float()
+        abs_err = (pred - y_norm).abs()
+        vol_mask = mask & ~is_surface
+        surf_mask = mask & is_surface
+        vol_loss = (abs_err * vol_mask.unsqueeze(-1)).sum() / vol_mask.sum().clamp(min=1)
+        surf_loss = (abs_err * surf_mask.unsqueeze(-1)).sum() / surf_mask.sum().clamp(min=1)
+        loss = vol_loss + cfg.surf_weight * surf_loss
 
         optimizer.zero_grad()
         loss.backward()
